aineon-enterprise/core/main.py
import os
import time
import asyncio
import json
from web3 import Web3
from dotenv import load_dotenv
from infrastructure.paymaster import PimlicoPaymaster
from strategies.jit_liquidity import JITManager

load_dotenv()

class AineonEngine:
    def __init__(self):
        # 1. Initialize Blockchain Connection
        self.w3 = Web3(Web3.HTTPProvider(os.getenv("ETH_RPC_URL")))
        self.paymaster = PimlicoPaymaster()
        self.jit_logic = JITManager()
        
        # 2. Load Contract
        # Note: In production, ABI should be loaded from a file
        self.contract_address = os.getenv("CONTRACT_ADDRESS") # Must be set after deployment
        self.account_address = os.getenv("WALLET_ADDRESS")
        self.private_key = os.getenv("PRIVATE_KEY") # Ensure this is in .env

    async def run(self):
        print(">> AINEON MAXIMIZED ENGINE STARTED")
        print(f">> Connected to Chain ID: {self.w3.eth.chain_id}")
        print(">> Waiting for Opportunities...")

        while True:
            try:
                # 1. SCAN: Get Price Data (Mocked for Logic flow)
                opportunity_found, trade_params = self.scan_market_opportunities()

                if opportunity_found:
                    print("[AI] Opportunity Detected. Calculating Gasless Route...")
                    
                    # 2. BUILD: Construct the Transaction (UserOp)
                    user_op = self.construct_flash_loan_tx(trade_params)
                    
                    # 3. SPONSOR: Ask Pimlico to Pay Gas
                    sponsored_op = self.paymaster.sponsor_transaction(user_op)
                    
                    if sponsored_op:
                        # 4. EXECUTE: Sign and Send Bundle
                        print(">>> EXECUTING SPONSORED FLASH LOAN <<<")
                        # In full implementation, sign user_op here
                        # self.paymaster.send_bundle(sponsored_op, signature)
                        print(">>> SUCCESS: Trade Submitted to Bundler.")
                    else:
                        print("[RISK] Paymaster rejected sponsorship (Unprofitable).")
                
            except Exception as e:
                print(f"[ERROR] Cycle: {e}")
            
            await asyncio.sleep(2)

    def scan_market_opportunities(self):
        # LOGIC: Check Uniswap vs Sushiswap spread
        # Return True if Spread > 0.5% (Configurable)
        return False, {} # Placeholder for safety

    def construct_flash_loan_tx(self, params):
        # LOGIC: Encode function data for AineonUltra.executeOperation
        # This creates the 'callData' for the UserOp
        return {
            "sender": self.account_address,
            "nonce": 0, # Should fetch real nonce
            "callData": "0x..." # Encoded ABI data
        }

if __name__ == "__main__":
    engine = AineonEngine()
    asyncio.run(engine.run())
