aineon-enterprise/requirements.txt
web3>=6.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
requests>=2.28.0
numpy>=1.21.0
